community(feat): further componentize resources section, localize css for member stories, resources, and events
