+++ src/components/stateless/Resources/Podcasts/Podcasts.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { ResourceLabel, ResourceTitle, ResourceDesc, ResourceImage, ResourceLink, PastResources } from '../ResourceCard';
+import { ResourceLabel, ResourceTitle, ResourceDesc, ResourceImage, ResourceLink, PastResources } from '../ResourceCard/ResourceCard';
 import podcastImage from '../../../../../src/resources/img/podcast-img.svg';
 import RedirectButton from '../../../state/RedirectButton/RedirectButton';
