+++ src/components/stateless/Resources/ResourceCard/ResourceCard.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './style.css';
+import '../style.css';
 
 class ResourceLabel extends Component {
   render() {
