+++ src/components/stateless/Resources/style.css
@@ -54,6 +54,13 @@
   margin-top: 2.5rem;
 }
 
+.resources-container {
+  height: 1230px;
+  background: #EEF1F6;
+  padding-top: 0.5rem;
+  position: relative;
+}
+
 .ResourceImage {
   grid-column: 4/9;
   grid-row: 1/12;
